address Ilya's feedback creating a base class for both cmdlets
